Allow building a shared library.
By default, Makefile does not set aunite_SOVERSION and the project
goes on building a static library. If aunit_SOVERSION is set in the
environment, build a shared library.

In order to allow two kind of builds to coexist, change the library
and object directories. These directories do not seem used outside the
project.